requirements.txt
yfinance==0.2.28
pandas==2.1.0
numpy==1.23.5
tqdm==4.66.1
scikit-learn==1.2.2
pytz==2023.3
finta==1.3
ta==0.10.2
keras==2.13.1
tensorflow==2.13.0
alpaca-py==0.10.0
matplotlib==3.7.2
tqdm==4.66.1
scipy==1.11.2
